agression og liv

+++ Evolution spil/Assets/scripts/mainMenu/LifePlus.cs
@@ -6,11 +6,11 @@ using UnityEngine.UI;
 public class LifePlus : MonoBehaviour
 {
     public Text value;
-    private int Life_nr = 1;
+    private float Life_nr = 1;
 
     void Start()
     {
-        Life_nr = PlayerPrefs.GetInt("Life");
+        Life_nr = PlayerPrefs.GetFloat("Life");
         value.text = "" + Life_nr;
     }
 
@@ -18,16 +18,16 @@ public class LifePlus : MonoBehaviour
     {
         if (Input.GetKeyDown(KeyCode.Mouse0))
         {
-            Life_nr = PlayerPrefs.GetInt("Life");
+            Life_nr = PlayerPrefs.GetFloat("Life");
             if (Life_nr <= 4)
             {
                 Life_nr = Life_nr + 1;
-                PlayerPrefs.SetInt("Life", Life_nr);
+                PlayerPrefs.SetFloat("Life", Life_nr);
             }
             else
             {
                 Life_nr = 1;
-                PlayerPrefs.SetInt("Life", Life_nr);
+                PlayerPrefs.SetFloat("Life", Life_nr);
             }
             value.text = "" + Life_nr;
             Debug.Log("Life Value = " + Life_nr);
